fix(frontend): remove drug effect and refactor ss-183
