+++ apps/frontend/src/libs/components/carousel/libs/constants/carausel-config.constant.ts
@@ -1,5 +1,4 @@
 const CAROUSEL_CONFIG = {
-	DRAG_MULTIPLIER: 0.15,
 	FRICTION: 0.96,
 	MIN_VELOCITY: 0.5,
 	SCROLL_SPEED: 0.5,
