--- apps/frontend/src/libs/components/carousel/libs/helpers/helpers.ts
@@ -1,1 +1,0 @@
-export { getCarouselParameters } from "./get-carousel-parameters.helper.js";
