+++ apps/frontend/src/libs/components/carousel/libs/hooks/hooks.ts
@@ -1,6 +1,4 @@
-export { useCarouselEvents } from "./use-carousel-events.hook.js";
 export { useCarouselMomentum } from "./use-carousel-momentum.hook.js";
-export { useCarouselMouseEvents } from "./use-carousel-mouse-events.hook.js";
+export { useMouseLeave } from "./use-carousel-mouse-leave.hook.js";
 export { useCarouselReference } from "./use-carousel-reference.hook.js";
-export { useCarouselStyles } from "./use-carousel-styles.hook.js";
 export { useCarouselWheelEvent } from "./use-carousel-wheel-event.hook.js";
