+++ apps/frontend/src/libs/components/carousel/libs/hooks/use-carousel-reference.hook.ts
@@ -4,18 +4,12 @@ import { type CarouselReference } from "../types/types.js";
 
 const useCarouselReference = (): CarouselReference => {
 	const element = useRef<HTMLDivElement>(null);
-	const isDragging = useRef(false);
-	const startX = useRef(0);
-	const scrollStart = useRef(0);
 	const velocity = useRef(0);
 	const momentumID = useRef<null | number>(null);
 
 	return {
 		element,
-		isDragging,
 		momentumID,
-		scrollStart,
-		startX,
 		velocity,
 	};
 };
