--- apps/frontend/src/libs/components/carousel/libs/hooks/use-carousel-styles.hook.ts
@@ -1,24 +1,0 @@
-import { combineClassNames } from "~/libs/helpers/helpers.js";
-
-import styles from "../../style.module.css";
-
-type CarouselStylesProperties = {
-	dragging: boolean;
-};
-
-const useCarouselStyles = ({
-	dragging,
-}: CarouselStylesProperties): {
-	carouselClassName: string;
-} => {
-	const carouselClassName = combineClassNames(
-		styles["carousel"],
-		dragging && styles["dragging"],
-	);
-
-	return {
-		carouselClassName,
-	};
-};
-
-export { useCarouselStyles };
