+++ apps/frontend/src/libs/components/carousel/libs/hooks/use-carousel-wheel-event.hook.ts
@@ -1,18 +1,13 @@
-import { useCallback } from "~/libs/hooks/hooks.js";
+import { useCallback, useEffect } from "~/libs/hooks/hooks.js";
 
 import { CAROUSEL_CONFIG } from "../constants/constants.js";
 import { type CarouselReference } from "../types/types.js";
 
-type CarouselWheelEventProperties = {
-	carouselReference: CarouselReference;
-	startMomentum: () => void;
-};
-
-const useCarouselWheelEvent = ({
-	carouselReference,
-	startMomentum,
-}: CarouselWheelEventProperties): ((event: WheelEvent) => void) => {
-	return useCallback(
+const useCarouselWheelEvent = (
+	carouselReference: CarouselReference,
+	startMomentum: () => void,
+): void => {
+	const handleWheelEvent = useCallback(
 		(event: WheelEvent): void => {
 			event.preventDefault();
 
@@ -27,6 +22,18 @@ const useCarouselWheelEvent = ({
 		},
 		[startMomentum, carouselReference],
 	);
+
+	useEffect(() => {
+		const element = carouselReference.element.current;
+
+		if (element) {
+			element.addEventListener("wheel", handleWheelEvent, { passive: false });
+
+			return (): void => {
+				element.removeEventListener("wheel", handleWheelEvent);
+			};
+		}
+	}, [handleWheelEvent, carouselReference.element]);
 };
 
 export { useCarouselWheelEvent };
