--- apps/frontend/src/libs/components/carousel/libs/types/carousel-parameters.type.ts
@@ -1,10 +1,0 @@
-type CarouselParameters = {
-	clientWidth: number;
-	element: HTMLDivElement | null;
-	maxScroll: number;
-	offsetLeft: number;
-	scrollLeft: number;
-	scrollWidth: number;
-};
-
-export { type CarouselParameters };
