+++ apps/frontend/src/libs/components/carousel/libs/types/carousel-reference.type.ts
@@ -1,9 +1,6 @@
 type CarouselReference = {
 	element: React.RefObject<HTMLDivElement | null>;
-	isDragging: React.RefObject<boolean>;
 	momentumID: React.RefObject<null | number>;
-	scrollStart: React.RefObject<number>;
-	startX: React.RefObject<number>;
 	velocity: React.RefObject<number>;
 };
