+++ apps/frontend/src/libs/components/carousel/libs/types/types.ts
@@ -1,2 +1,1 @@
-export { type CarouselParameters } from "./carousel-parameters.type.js";
 export { type CarouselReference } from "./carousel-reference.type.js";
