+++ apps/frontend/src/libs/components/carousel/style.module.css
@@ -12,13 +12,7 @@
 	max-height: 100%;
 	overflow-x: auto;
 	scroll-behavior: auto;
-	cursor: grab;
 	transition: transform 0.1s ease-out;
-}
-
-.carousel.dragging {
-	cursor: grabbing;
-	transition: transform 1s ease-out;
 }
 
 .carousel-item {
