+++ apps/frontend/src/libs/hooks/hooks.ts
@@ -1,6 +1,6 @@
 export { useAppDispatch } from "./use-app-dispatch/use-app-dispatch.hook.js";
 export { useAppForm } from "./use-app-form/use-app-form.hook.js";
 export { useAppSelector } from "./use-app-selector/use-app-selector.hook.js";
-export { useCallback, useEffect, useRef, useState } from "react";
+export { useCallback, useEffect, useRef } from "react";
 export { useController as useFormController } from "react-hook-form";
 export { useLocation } from "react-router";
